Fix display hanging for 1 second when timepulse occurs during UART receive
It is possible for a timepulse to occur during NMEA message output from
the receiver if not locked to GPS time. This meant that a timepulse could
occur before the next display value had been prepared, and the existing
value would be sent again. This appeared as the clock pausing for two
seconds on the same display.

I've added handling of repeated timepulses so the display updates correctly
if the latest time from a NMEA message been loaded into the display buffer
yet.

This problem likely goes away if the timepulse is configured to only occur
when there is a GPS lock, since the asynchronous timepulse and GPS message
output seems to be the cause of this overlap. The overlap was happening
every 5 seconds once the UART and timepulse had got into a rhythm