Add a script to decode digit data sent to MAX7221 chip
Given the flipped wiring used for common-anode segments, plus digits not
being in order in the v1 hardware, looking at bytes sent to the MAX7221
using a logic analyser gives no useful information. This script decodes
sequences of 8 bytes sent to the the displayed back to the characters that
are being displayed.

This is useful for relating what was sent to other events while probing.