Add OpenOCD config to allow attaching for debug and remote resetting
This build isn't currently set up for debugging with gdb, but this allows
simple operations and viewing memory. Debuggin with gdb requires an elf
build output, which sdcc can do, but it's a little fiddly:

 - Function-static variables aren't supported and cause the linker to segfault
 - stm8flash doesn't support elf format output to upload
 - sdobjcopy doesn't support elf format, so can't convert to ihx/bin

This means to get a debug build currently, the build has to be run twice
with the --debug flag: first for a hex file output, then again for an elf
output. This isn't ideal, but seems to work in a pinch.