Convert main display and button handling into state machines
This implements timezone incrementing in a similar style to avr-doomclock:
pressing the button once shows the current UTC offset, and holding the
button increments the timezone at a comfortable pace until released.

There may be bugs buried in here and missing comments. This is recovered
from source comments in main.asm which I happened to have open after I
closed the buffer containing the completed code when setting up main.c in
an older state for more atomic commits :|